+++ test/api-suite.js
@@ -355,6 +355,40 @@ console.log('\nTest 17: Stereo Rendering');
     assert(left[100] === right[100], 'Stereo channels are identical');
 }
 
+// Test 18: AudioBufferSourceNode resamples mismatched buffer sample rates
+console.log('\nTest 18: AudioBufferSourceNode Resampling');
+{
+    const inputSampleRate = 1000;
+    const outputSampleRate = 2000;
+    const inputLength = 1000;
+    const outputLength = Math.ceil((inputLength * outputSampleRate) / inputSampleRate);
+    const ctx = new OfflineAudioContext({
+        numberOfChannels: 1,
+        length: outputLength,
+        sampleRate: outputSampleRate
+    });
+
+    const buffer = ctx.createBuffer(1, inputLength, inputSampleRate);
+    buffer.getChannelData(0).fill(1);
+
+    const source = ctx.createBufferSource();
+    source.buffer = buffer;
+    source.connect(ctx.destination);
+    source.start(0);
+
+    const rendered = await ctx.startRendering();
+    const data = rendered.getChannelData(0);
+    const earlyAverage =
+        data.subarray(100, 500).reduce((sum, value) => sum + value, 0) / 400;
+    const lateAverage =
+        data.subarray(1400, 1800).reduce((sum, value) => sum + value, 0) / 400;
+
+    assert(data.length === outputLength, 'Output length matches resampled duration');
+    assert(rendered.sampleRate === outputSampleRate, 'Output sample rate is context sample rate');
+    assert(earlyAverage > 0.9, 'Resampled buffer starts near full scale');
+    assert(lateAverage > 0.9, 'Resampled buffer remains active for full duration');
+}
+
 // Summary
 console.log(`\n${'='.repeat(50)}`);
 console.log(`Test Results: ${passed} passed, ${failed} failed`);
